rename TensorConverter -> TensorSerializer
