+++ tests/converter/test_utils.py
@@ -1,8 +1,8 @@
-"""Test utils functions for the converter module."""
+"""Test utils functions for the serialization module."""
 
 import pytest
 
-from tensorshare.converter.utils import (
+from tensorshare.serialization.utils import (
     convert_flax_to_safetensors,
     convert_numpy_to_safetensors,
     convert_paddle_to_safetensors,
